+++ antarest/study/storage/variantstudy/business/matrix_constants/__init__.py
@@ -1,1 +1,1 @@
-from . import hydro, link, prepro, st_storage, thermals
+from . import binding_constraint, hydro, link, prepro, st_storage, thermals