+++ antarest/study/storage/variantstudy/model/command/create_binding_constraint.py
@@ -1,12 +1,14 @@
+from abc import ABCMeta
 from typing import Any, Dict, List, Optional, Tuple, Union, cast
 
-from pydantic import validator
+import numpy as np
+from pydantic import Field, validator
 
-from antarest.core.utils.utils import assert_this
 from antarest.matrixstore.model import MatrixData
 from antarest.study.storage.rawstudy.model.filesystem.config.binding_constraint import BindingConstraintFrequency
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig, transform_name_to_id
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
+from antarest.study.storage.variantstudy.business.matrix_constants_generator import GeneratorMatrixConstants
 from antarest.study.storage.variantstudy.business.utils import strip_matrix_protocol, validate_matrix
 from antarest.study.storage.variantstudy.business.utils_binding_constraint import (
     apply_binding_constraint,
@@ -20,34 +22,115 @@ from antarest.study.storage.variantstudy.model.command.common import (
 from antarest.study.storage.variantstudy.model.command.icommand import MATCH_SIGNATURE_SEPARATOR, ICommand
 from antarest.study.storage.variantstudy.model.model import CommandDTO
 
+__all__ = ("AbstractBindingConstraintCommand", "CreateBindingConstraint", "check_matrix_values")
 
-class CreateBindingConstraint(ICommand):
-    name: str
+MatrixType = List[List[MatrixData]]
+
+
+def check_matrix_values(time_step: BindingConstraintFrequency, values: MatrixType) -> None:
+    """
+    Check the binding constraint's matrix values for the specified time step.
+
+    Args:
+        time_step: The frequency of the binding constraint: "hourly", "daily" or "weekly".
+        values: The binding constraint's 2nd member matrix.
+
+    Raises:
+        ValueError:
+            If the matrix shape does not match the expected shape for the given time step.
+            If the matrix values contain NaN (Not-a-Number).
+    """
+    shapes = {
+        BindingConstraintFrequency.HOURLY: (8760, 3),
+        BindingConstraintFrequency.DAILY: (365, 3),
+        BindingConstraintFrequency.WEEKLY: (52, 3),
+    }
+    # Check the matrix values and create the corresponding matrix link
+    array = np.array(values, dtype=np.float64)
+    if array.shape != shapes[time_step]:
+        raise ValueError(f"Invalid matrix shape {array.shape}, expected {shapes[time_step]}")
+    if np.isnan(array).any():
+        raise ValueError("Matrix values cannot contain NaN")
+
+
+class AbstractBindingConstraintCommand(ICommand, metaclass=ABCMeta):
+    """
+    Abstract class for binding constraint commands.
+    """
+
+    # todo: add the `name` attribute because it should also be updated
     enabled: bool = True
     time_step: BindingConstraintFrequency
     operator: BindingConstraintOperator
     coeffs: Dict[str, List[float]]
-    values: Optional[Union[List[List[MatrixData]], str]] = None
+    values: Optional[Union[MatrixType, str]] = Field(None, description="2nd member matrix")
     filter_year_by_year: Optional[str] = None
     filter_synthesis: Optional[str] = None
     comments: Optional[str] = None
 
-    def __init__(self, **data: Any) -> None:
-        super().__init__(
-            command_name=CommandName.CREATE_BINDING_CONSTRAINT,
-            version=1,
-            **data,
+    def to_dto(self) -> CommandDTO:
+        args = {
+            "enabled": self.enabled,
+            "time_step": self.time_step.value,
+            "operator": self.operator.value,
+            "coeffs": self.coeffs,
+            "comments": self.comments,
+            "filter_year_by_year": self.filter_year_by_year,
+            "filter_synthesis": self.filter_synthesis,
+        }
+        if self.values is not None:
+            args["values"] = strip_matrix_protocol(self.values)
+        return CommandDTO(
+            action=self.command_name.value,
+            args=args,
         )
+
+    def get_inner_matrices(self) -> List[str]:
+        if self.values is not None:
+            if not isinstance(self.values, str):  # pragma: no cover
+                raise TypeError(repr(self.values))
+            return [strip_matrix_protocol(self.values)]
+        return []
+
+
+class CreateBindingConstraint(AbstractBindingConstraintCommand):
+    """
+    Command used to create a binding constraint.
+    """
+
+    command_name: CommandName = CommandName.CREATE_BINDING_CONSTRAINT
+    version: int = 1
+
+    # Properties of the `CREATE_BINDING_CONSTRAINT` command:
+    name: str
 
     @validator("values", always=True)
     def validate_series(
-        cls, v: Optional[Union[List[List[MatrixData]], str]], values: Any
-    ) -> Optional[Union[List[List[MatrixData]], str]]:
+        cls,
+        v: Optional[Union[MatrixType, str]],
+        values: Dict[str, Any],
+    ) -> Optional[Union[MatrixType, str]]:
+        constants: GeneratorMatrixConstants
+        constants = values["command_context"].generator_matrix_constants
+        time_step = values["time_step"]
         if v is None:
-            v = values["command_context"].generator_matrix_constants.get_null_matrix()
-            return v
-        else:
+            # Use an already-registered default matrix
+            methods = {
+                BindingConstraintFrequency.HOURLY: constants.get_binding_constraint_hourly,
+                BindingConstraintFrequency.DAILY: constants.get_binding_constraint_daily,
+                BindingConstraintFrequency.WEEKLY: constants.get_binding_constraint_weekly,
+            }
+            method = methods[time_step]
+            return method()
+        if isinstance(v, str):
+            # Check the matrix link
             return validate_matrix(v, values)
+        if isinstance(v, list):
+            check_matrix_values(time_step, v)
+            return validate_matrix(v, values)
+        # Invalid datatype
+        # pragma: no cover
+        raise TypeError(repr(v))
 
     def _apply_config(self, study_data_config: FileStudyTreeConfig) -> Tuple[CommandOutput, Dict[str, Any]]:
         bd_id = transform_name_to_id(self.name)
@@ -55,7 +138,6 @@ class CreateBindingConstraint(ICommand):
         return CommandOutput(status=True), {}
 
     def _apply(self, study_data: FileStudy) -> CommandOutput:
-        assert_this(isinstance(self.values, str))
         binding_constraints = study_data.tree.get(["input", "bindingconstraints", "bindingconstraints"])
         new_key = len(binding_constraints.keys())
         bd_id = transform_name_to_id(self.name)
@@ -76,20 +158,9 @@ class CreateBindingConstraint(ICommand):
         )
 
     def to_dto(self) -> CommandDTO:
-        return CommandDTO(
-            action=CommandName.CREATE_BINDING_CONSTRAINT.value,
-            args={
-                "name": self.name,
-                "enabled": self.enabled,
-                "time_step": self.time_step.value,
-                "operator": self.operator.value,
-                "coeffs": self.coeffs,
-                "values": strip_matrix_protocol(self.values),
-                "comments": self.comments,
-                "filter_year_by_year": self.filter_year_by_year,
-                "filter_synthesis": self.filter_synthesis,
-            },
-        )
+        dto = super().to_dto()
+        dto.args["name"] = self.name  # type: ignore
+        return dto
 
     def match_signature(self) -> str:
         return str(self.command_name.value + MATCH_SIGNATURE_SEPARATOR + self.name)
@@ -129,9 +200,3 @@ class CreateBindingConstraint(ICommand):
                 command_context=other.command_context,
             )
         ]
-
-    def get_inner_matrices(self) -> List[str]:
-        if self.values is not None:
-            assert_this(isinstance(self.values, str))
-            return [strip_matrix_protocol(self.values)]
-        return []